feat(Account/Vip): change logger category to `module` and add reload commands

+++ src/server/scripts/Vip/Vip_SC.cpp
@@ -34,7 +34,7 @@ public:
         if (!sVip->IsEnable() || !sVip->IsVip(player))
             return;
 
-        amount *= sVip->GetRateForPlayer(player, VipRate::XP);
+        amount *= static_cast<uint32>(sVip->GetRateForPlayer(player, VipRate::XP));
     }
 
     void OnGiveHonorPoints(Player* player, float& points, Unit* /*victim*/) override
@@ -50,7 +50,7 @@ public:
         if (!sVip->IsEnable() || !sVip->IsVip(player))
             return true;
 
-        standing *= sVip->GetRateForPlayer(player, VipRate::Reputation);
+        standing *= static_cast<int32>(sVip->GetRateForPlayer(player, VipRate::Reputation));
         return true;
     }
